--- a/src/components/auth/SignInButton/style.module.scss
+++ b/src/components/auth/SignInButton/style.module.scss
@@ -1,6 +1,5 @@
 .link {
   color: var(--theme-text-on-background-1);
-  font-family: 'DM Sans', sans-serif;
   font-size: 16px;
   font-weight: 400;
   line-height: 150%;
@@ -19,7 +18,6 @@
   color: var(--theme-background);
   display: flex;
 
-  font-family: 'DM Sans', sans-serif;
   font-size: 20px;
   font-weight: 700;
   height: 2.25rem;
@@ -42,7 +40,6 @@
   color: var(--theme-background);
   display: flex;
 
-  font-family: 'DM Sans', sans-serif;
   font-size: 20px;
   font-weight: 700;
   height: 2.25rem;