--- a/src/pages/_app.tsx
+++ b/src/pages/_app.tsx
@@ -11,19 +11,30 @@ import { CookieType } from '@/lib/types/enums';
 import { NextPageContext } from 'next';
 import { ThemeProvider } from 'next-themes';
 import type { AppProps } from 'next/app';
+// eslint-disable-next-line camelcase
+import { DM_Sans } from 'next/font/google';
 import { ToastContainer } from 'react-toastify';
 
 interface InitialPropInterface {
   user: PrivateProfile;
 }
+const dmSans = DM_Sans({ subsets: ['latin'], weight: ['400', '500', '700'] });
+
 export default function MyApp({ Component, pageProps }: AppProps<InitialPropInterface>) {
   return (
-    <ThemeProvider>
-      <ToastContainer />
-      <PageLayout user={pageProps?.user}>
-        <Component {...pageProps} />
-      </PageLayout>
-    </ThemeProvider>
+    <>
+      <style jsx global>{`
+        * {
+          font-family: ${dmSans.style.fontFamily}, sans-serif;
+        }
+      `}</style>
+      <ThemeProvider>
+        <ToastContainer />
+        <PageLayout user={pageProps?.user}>
+          <Component {...pageProps} />
+        </PageLayout>
+      </ThemeProvider>
+    </>
   );
 }