--- a/src/styles/globals.scss
+++ b/src/styles/globals.scss
@@ -1,5 +1,4 @@
 @use './vars.scss' as vars;
-@import url('https://fonts.googleapis.com/css2?family=DM+Sans:ital,wght@0,400;0,500;0,700;1,400;1,500;1,700&display=swap');
 
 html,
 body {
@@ -7,10 +6,6 @@ body {
   line-height: 1.5;
   margin: 0;
   padding: 0;
-}
-
-* {
-  font-family: 'DM Sans', sans-serif;
 }
 
 main {