# Description

On initial or full page loads of the Membership Portal (disable caching), fetching font resources files for DM Sans takes 

`579ms` out of `720ms` for the total page load (80% of page load time)
![Screenshot 2023-04-12 at 11 49 45 PM](https://user-images.githubusercontent.com/33165426/231677774-a4c0e559-80ac-4d26-b912-f6fca56b1e2d.jpg)

`294ms` out of `343ms` for the total page load (85% of page load time)
![Screenshot 2023-04-12 at 11 50 21 PM](https://user-images.githubusercontent.com/33165426/231677797-f5160da0-2331-4c0d-bd7a-fb76b48fe44c.jpg)

The solution to this significant data fetch is to self-host our fonts in the repo and define the font-face "DM Sans" to pull from the file that stores the font-face configuration. Next.js offers a better solution than doing it from scratch in only a few lines of code that offers the exact same solution but we don't actually have to perform this full setup - it will automatically happen at build time when the project is compiled.

With the change proposed in this PR, font loading now takes `~60ms` instead of `~500ms` with a total page load time of under `200ms` reducing the load time by nearly 90%.

See [Font Optimization](https://nextjs.org/docs/basic-features/font-optimization) for official Next.js documentation.

## Changes

- Switches to Next.js recommended font import method.

# Type of Change

- [x] Bug Fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as
      expected)
- [ ] Logistics Change (A change to a README, description, or dev workflow setup like
      linting/formatting)
- [ ] Continuous Integration Change (Related to deployment steps or continuous integration
      workflows)
- [ ] Other: (Fill In) <!-- Edit this type of change if you select this -->

# Testing

I have tested that my changes fully resolve the linked issue ...

- [ ] locally on Desktop.
- [ ] locally on mobile - use https://ngrok.io to get a copy on a mobile device
- [ ] on the live deployment preview on Desktop.
- [ ] on the live deployment preview on Mobile.
- [ ] I have run and passed all new and existing Cypress tests. Add screenshots below.

# Checklist

- [ ] I have performed a self-review of my own code.
- [ ] I have followed the style guidelines of this project.
- [ ] I have documented my code's `src/lib` functions and commented hard to understand areas
      anywhere else.
- [ ] My changes produce no new warnings.

# Screenshots

Please include a screenshot of your Cypress testing suite passing successfully.

If you made any visual changes to the website, please include relevant screenshots below.